refactor: factory moved outside checks
